Add loading effects, change file structure
