--- .eslintcache
@@ -1,1 +1,0 @@
-[{"/home/lucas/programming/microverse/react/react-capstone/src/index.js":"1","/home/lucas/programming/microverse/react/react-capstone/src/Routes.js":"2","/home/lucas/programming/microverse/react/react-capstone/src/store/store.js":"3","/home/lucas/programming/microverse/react/react-capstone/src/pages/Home.js":"4","/home/lucas/programming/microverse/react/react-capstone/src/pages/Details.js":"5","/home/lucas/programming/microverse/react/react-capstone/src/reducer/index.js":"6","/home/lucas/programming/microverse/react/react-capstone/src/components/StockList.js":"7","/home/lucas/programming/microverse/react/react-capstone/src/components/Navbar.js":"8","/home/lucas/programming/microverse/react/react-capstone/src/API/api.js":"9","/home/lucas/programming/microverse/react/react-capstone/src/reducer/filter.js":"10","/home/lucas/programming/microverse/react/react-capstone/src/reducer/profile.js":"11","/home/lucas/programming/microverse/react/react-capstone/src/components/Filter.js":"12","/home/lucas/programming/microverse/react/react-capstone/src/components/StockCard.js":"13","/home/lucas/programming/microverse/react/react-capstone/src/actions/index.js":"14","/home/lucas/programming/microverse/react/react-capstone/src/utils/index.js":"15","/home/lucas/programming/microverse/react/react-capstone/src/constants/index.js":"16","/home/lucas/programming/microverse/react/react-capstone/src/constants/initial.js":"17","/home/lucas/programming/microverse/react/react-capstone/src/store/async.js":"18"},{"size":424,"mtime":1613751163746,"results":"19","hashOfConfig":"20"},{"size":393,"mtime":1613563116299,"results":"21","hashOfConfig":"20"},{"size":368,"mtime":1613751177850,"results":"22","hashOfConfig":"20"},{"size":232,"mtime":1613659990867,"results":"23","hashOfConfig":"20"},{"size":13293,"mtime":1613648124922,"results":"24","hashOfConfig":"20"},{"size":218,"mtime":1613523066068,"results":"25","hashOfConfig":"20"},{"size":1527,"mtime":1613762405011,"results":"26","hashOfConfig":"20"},{"size":824,"mtime":1613679437315,"results":"27","hashOfConfig":"20"},{"size":1023,"mtime":1613752701346,"results":"28","hashOfConfig":"20"},{"size":437,"mtime":1613488427974,"results":"29","hashOfConfig":"20"},{"size":309,"mtime":1613752054310,"results":"30","hashOfConfig":"20"},{"size":4696,"mtime":1613681650085,"results":"31","hashOfConfig":"20"},{"size":3897,"mtime":1613659795651,"results":"32","hashOfConfig":"20"},{"size":177,"mtime":1613488302526,"results":"33","hashOfConfig":"20"},{"size":745,"mtime":1613756500502,"results":"34","hashOfConfig":"20"},{"size":2030,"mtime":1613511454606,"results":"35","hashOfConfig":"20"},{"size":1722052,"mtime":1613395360943,"results":"36","hashOfConfig":"20"},{"size":210,"mtime":1613751004858,"results":"37","hashOfConfig":"20"},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1rfntn1",{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"42"},{"filePath":"43","messages":"44","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"45","messages":"46","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"42"},{"filePath":"47","messages":"48","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"42"},{"filePath":"49","messages":"50","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"42"},{"filePath":"51","messages":"52","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"53","messages":"54","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"42"},{"filePath":"55","messages":"56","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"57","messages":"58","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"42"},{"filePath":"59","messages":"60","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"61","messages":"62","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"42"},{"filePath":"63","messages":"64","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"42"},{"filePath":"65","messages":"66","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"42"},{"filePath":"67","messages":"68","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"42"},{"filePath":"69","messages":"70","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"42"},{"filePath":"71","messages":"72","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"42"},{"filePath":"73","messages":"74","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/lucas/programming/microverse/react/react-capstone/src/index.js",[],"/home/lucas/programming/microverse/react/react-capstone/src/Routes.js",[],["75","76","77","78","79","80"],"/home/lucas/programming/microverse/react/react-capstone/src/store/store.js",[],"/home/lucas/programming/microverse/react/react-capstone/src/pages/Home.js",[],"/home/lucas/programming/microverse/react/react-capstone/src/pages/Details.js",[],"/home/lucas/programming/microverse/react/react-capstone/src/reducer/index.js",[],"/home/lucas/programming/microverse/react/react-capstone/src/components/StockList.js",["81"],"/home/lucas/programming/microverse/react/react-capstone/src/components/Navbar.js",[],"/home/lucas/programming/microverse/react/react-capstone/src/API/api.js",[],"/home/lucas/programming/microverse/react/react-capstone/src/reducer/filter.js",[],"/home/lucas/programming/microverse/react/react-capstone/src/reducer/profile.js",[],"/home/lucas/programming/microverse/react/react-capstone/src/components/Filter.js",[],"/home/lucas/programming/microverse/react/react-capstone/src/components/StockCard.js",[],"/home/lucas/programming/microverse/react/react-capstone/src/actions/index.js",[],"/home/lucas/programming/microverse/react/react-capstone/src/utils/index.js",[],"/home/lucas/programming/microverse/react/react-capstone/src/constants/index.js",[],"/home/lucas/programming/microverse/react/react-capstone/src/constants/initial.js",[],"/home/lucas/programming/microverse/react/react-capstone/src/store/async.js",[],{"ruleId":"82","replacedBy":"83"},{"ruleId":"84","replacedBy":"85"},{"ruleId":"86","replacedBy":"87"},{"ruleId":"88","replacedBy":"89"},{"ruleId":"90","replacedBy":"91"},{"ruleId":"92","replacedBy":"93"},{"ruleId":"94","severity":1,"message":"95","line":19,"column":3,"nodeType":"96","messageId":"97","endLine":19,"endColumn":14},"lines-around-directive",["98"],"no-spaced-func",["99"],"global-require",[],"no-buffer-constructor",[],"no-new-require",[],"no-path-concat",[],"no-console","Unexpected console statement.","MemberExpression","unexpected","padding-line-between-statements","func-call-spacing"]
