+++ package.json
@@ -13,6 +13,7 @@
     "prop-types": "^15.7.2",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "react-loading-components": "^1.1.12",
     "react-redux": "^7.1.3",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
