+++ src/API/api.js
@@ -1,4 +1,4 @@
-const API_KEY = '&apikey=f25597924d4529b7e1d2e609c841a5aa';
+const API_KEY = 'apikey=f25597924d4529b7e1d2e609c841a5aa';
 const PROFILE_URL = 'https://financialmodelingprep.com/api/v3/profile/';
 const LIST_URL = 'https://financialmodelingprep.com/api/v3/stock-screener?limit=500';
 const QUOTE_URL = 'https://financialmodelingprep.com/api/v3/quote/';
@@ -15,7 +15,7 @@ const profile = async company => {
 
 const quote = async symbol => {
   try {
-    const response = await fetch(`${QUOTE_URL}${symbol}${API_KEY}`, { mode: 'cors' });
+    const response = await fetch(`${QUOTE_URL}${symbol}?${API_KEY}`, { mode: 'cors' });
     const result = await response.json();
     return result[0];
   } catch (err) {
@@ -25,7 +25,7 @@ const quote = async symbol => {
 
 const fetchStocks = async () => {
   try {
-    const response = await fetch(`${LIST_URL}${API_KEY}`, { mode: 'cors' });
+    const response = await fetch(`${LIST_URL}&${API_KEY}`, { mode: 'cors' });
     const result = await response.json();
     return result;
   } catch (err) {
