+++ src/__tests__/stocklist/StockList.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
-import StockList from '../../components/StockList';
+import StockList from '../../pages/home/components/StockList';
 import rootReducer from '../../reducer/index';
 
 it('renders correctly when store is provided', () => {
