--- src/pages/Details.js
@@ -1,313 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import { useParams } from 'react-router';
-import { useHistory } from 'react-router-dom';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faLongArrowAltLeft } from '@fortawesome/free-solid-svg-icons';
-import Navbar from '../components/Navbar';
-import { quote } from '../API/api';
-
-const Details = () => {
-  const { symbol } = useParams();
-  const history = useHistory();
-  const [state, setState] = useState({
-    symbol: 'AAPL',
-    name: 'Apple Inc.',
-    price: 120.96000000,
-    changesPercentage: 0.07000000,
-    change: 0.08000000,
-    dayLow: 110.89000000,
-    dayHigh: 123.70000000,
-    yearHigh: 137.98000000,
-    yearLow: 52.76750000,
-    marketCap: 2068718419968.00000000,
-    priceAvg50: 112.15875000,
-    priceAvg200: 85.41895000,
-    volume: 332607163,
-    avgVolume: 165778904,
-    exchange: 'NASDAQ',
-    open: 120.07000000,
-    previousClose: 120.88000000,
-    eps: 3.29600000,
-    pe: 36.69902800,
-    earningsAnnouncement: '2020-07-30T16:30:00.000+0000',
-    sharesOutstanding: 17102500165,
-    timestamp: 1599435459,
-  });
-
-  useEffect(async () => {
-    const res = await quote(symbol);
-    if (res) {
-      setState(res);
-    }
-  }, []);
-
-  return (
-    <>
-      <Navbar />
-      <button
-        className="no-outline"
-        type="button"
-        onClick={() => history.goBack()}
-        style={{
-          position: 'absolute',
-          backgroundColor: '#17A2B8',
-          width: '5%',
-          height: '7%',
-          top: '150px',
-          border: 'none',
-          borderBottomRightRadius: '10px',
-          borderTopRightRadius: '10px',
-          color: 'var(--light)',
-          fontFamily: 'Alfa Slab One, cursive',
-          display: 'flex',
-          alignItems: 'center',
-          paddingLeft: '1.5%',
-        }}
-      >
-        <FontAwesomeIcon icon={faLongArrowAltLeft} size="3x" />
-      </button>
-      <div className="container d-flex flex-row justify-content-between" style={{ marginTop: 30, height: '75vh', width: '95%' }}>
-        <div style={{
-          height: '100%', width: '69%', background: 'rgba(108,117,125,0.16)', borderRadius: 20, padding: 20,
-        }}
-        >
-          <div
-            className="d-flex flex-row justify-content-between align-items-center"
-            style={{
-              height: '15%', width: '100%', fontSize: 40, fontFamily: '"Alfa Slab One", cursive', borderStyle: 'none', borderBottomStyle: 'solid',
-            }}
-          >
-            <span style={{ fontSize: 50, height: 60 }}>{state.symbol}</span>
-          </div>
-          <div style={{ height: '5%', marginBottom: 20 }}><span style={{ color: 'var(--blue)', fontFamily: '"Alfa Slab One", cursive' }}>{state.exchange}</span></div>
-          <div className="d-flex flex-row" style={{ height: '80%' }}>
-            <div style={{
-              background: 'var(--dark)', width: '61%', height: '90%', borderRadius: 20, padding: 15,
-            }}
-            >
-              <div className="d-flex flex-row justify-content-between">
-                <span style={{
-                  fontSize: 20, fontFamily: 'Alata, sans-serif', fontWeight: 'bold', marginRight: 0, color: 'var(--light)',
-                }}
-                >
-                  MktCap:
-                </span>
-                <span
-                  className="d-lg-flex align-items-lg-end"
-                  style={{
-                    fontSize: 18, fontFamily: 'Alata, sans-serif', color: 'var(--success)', borderTopColor: 'var(--green)',
-                  }}
-                >
-                  {`U$ ${state.marketCap.toFixed(2)}`}
-                </span>
-              </div>
-              <div className="d-flex flex-row justify-content-between">
-                <span style={{
-                  fontSize: 20, fontFamily: 'Alata, sans-serif', fontWeight: 'bold', marginRight: 20, color: 'var(--light)',
-                }}
-                >
-                  SharesOut:
-                </span>
-                <span style={{ fontSize: 18, fontFamily: 'Alata, sans-serif', color: 'var(--blue)' }}>{state.sharesOutstanding}</span>
-              </div>
-              <div className="d-flex flex-row justify-content-center" style={{ borderTopStyle: 'solid', borderTopColor: 'var(--light)' }}>
-                <span style={{
-                  fontSize: 20, fontFamily: 'Alata, sans-serif', fontWeight: 'bold', marginRight: 20, marginTop: 15, marginBottom: 15, fontStyle: 'italic', color: 'var(--light)',
-                }}
-                >
-                  Transactions
-                </span>
-              </div>
-              <div className="d-flex flex-row justify-content-between">
-                <span style={{
-                  fontSize: 20, fontFamily: 'Alata, sans-serif', fontWeight: 'bold', marginRight: 20, color: 'var(--light)',
-                }}
-                >
-                  Volume:
-                </span>
-                <span style={{ fontSize: 18, fontFamily: 'Alata, sans-serif', color: 'var(--blue)' }}>{state.volume}</span>
-              </div>
-              <div className="d-flex flex-row justify-content-between">
-                <span style={{
-                  fontSize: 20, fontFamily: 'Alata, sans-serif', fontWeight: 'bold', marginRight: 20, color: 'var(--light)',
-                }}
-                >
-                  AvgVolume:
-                </span>
-                <span style={{ fontSize: 18, fontFamily: 'Alata, sans-serif', color: 'var(--blue)' }}>{state.avgVolume}</span>
-              </div>
-              <div className="d-flex flex-row justify-content-between">
-                <span style={{
-                  fontSize: 20, fontFamily: 'Alata, sans-serif', fontWeight: 'bold', marginRight: 20, color: 'var(--light)',
-                }}
-                >
-                  PriceAvg50:
-                </span>
-                <span style={{ fontSize: 18, fontFamily: 'Alata, sans-serif', color: 'var(--success)' }}>{`U$ ${state.priceAvg50.toFixed(2)}`}</span>
-              </div>
-              <div className="d-flex flex-row justify-content-between">
-                <span style={{
-                  fontSize: 20, fontFamily: 'Alata, sans-serif', fontWeight: 'bold', marginRight: 20, color: 'var(--light)',
-                }}
-                >
-                  PriceAvg200:
-                </span>
-                <span style={{ fontSize: 18, fontFamily: 'Alata, sans-serif', color: 'var(--success)' }}>{`U$ ${state.priceAvg200.toFixed(2)}`}</span>
-              </div>
-            </div>
-            <div style={{
-              height: '100%', width: '39%', padding: 15, paddingTop: 0, paddingLeft: 25,
-            }}
-            >
-              <div className="d-flex flex-row justify-content-start align-items-center">
-                <span style={{
-                  fontFamily: 'Alata, sans-serif', fontSize: 40, height: 50, margin: '-5px',
-                }}
-                >
-                  Price
-                </span>
-              </div>
-              <div className="d-flex flex-row justify-content-start align-items-center" style={{ height: '20%' }}>
-                <div className="d-flex" style={{ height: 40, borderTopStyle: 'none' }}>
-                  <span
-                    className="d-flex flex-row justify-content-center align-items-center"
-                    style={{
-                      fontFamily: 'Aleo, serif', color: 'rgb(255,214,0)', fontSize: 38, fontWeight: 'bold', textShadow: '1px 1px 2px rgb(0,0,0)',
-                    }}
-                  >
-                    {`U$ ${state.price.toFixed(2)}`}
-                  </span>
-                </div>
-              </div>
-              <div className="d-flex flex-row justify-content-start align-items-center" style={{ marginTop: 20 }}>
-                <span style={{
-                  fontFamily: 'Alata, sans-serif', fontSize: 25, height: 30, margin: '-5px',
-                }}
-                >
-                  Open
-                </span>
-              </div>
-              <div className="d-flex flex-row justify-content-start align-items-center" style={{ height: '20%' }}>
-                <div className="d-flex" style={{ height: 40, borderTopStyle: 'none' }}>
-                  <span
-                    className="d-flex flex-row justify-content-center align-items-center"
-                    style={{
-                      fontFamily: 'Alata, sans-serif', color: 'var(--green)', fontSize: 28, fontWeight: 'bold',
-                    }}
-                  >
-                    {`U$ ${state.open.toFixed(2)}`}
-                  </span>
-                </div>
-              </div>
-              <div className="d-flex flex-row justify-content-start align-items-center">
-                <span style={{
-                  fontFamily: 'Alata, sans-serif', fontSize: 25, height: 30, margin: '-5px',
-                }}
-                >
-                  Previous Closed
-                </span>
-              </div>
-              <div className="d-flex flex-row justify-content-start align-items-center" style={{ height: '20%' }}>
-                <div className="d-flex" style={{ height: 40, borderTopStyle: 'none' }}>
-                  <span
-                    className="d-flex flex-row justify-content-center align-items-center"
-                    style={{
-                      fontFamily: 'Alata, sans-serif', color: 'var(--danger)', fontSize: 28, fontWeight: 'bold',
-                    }}
-                  >
-                    {`U$ ${state.previousClose.toFixed(2)}`}
-                  </span>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div style={{
-          height: '100%', width: '30%', background: 'var(--dark)', borderRadius: 20, padding: 20, paddingTop: 10,
-        }}
-        >
-          <div className="d-flex justify-content-center" style={{ marginBottom: 15 }}>
-            <span style={{
-              fontSize: 36, fontFamily: 'Alata, sans-serif', color: 'var(--light)', fontWeight: 'bold', fontStyle: 'italic',
-            }}
-            >
-              Quote
-            </span>
-          </div>
-          <div className="d-flex flex-column align-items-end">
-            <div style={{
-              borderBottomStyle: 'solid', borderBottomColor: 'var(--light)', height: 30, width: '100%',
-            }}
-            >
-              <span style={{ color: 'var(--light)', fontSize: 20, fontFamily: 'Alata, sans-serif' }}>Change (%)</span>
-            </div>
-            <span style={{
-              color: 'var(--yellow)', fontSize: 28, fontFamily: 'Alata, sans-serif', fontWeight: 'bold', margin: '-5px',
-            }}
-            >
-              {state.changesPercentage}
-            </span>
-          </div>
-          <div className="d-flex flex-column align-items-end">
-            <div style={{
-              borderBottomStyle: 'solid', borderBottomColor: 'var(--light)', height: 30, width: '100%',
-            }}
-            >
-              <span style={{ color: 'var(--light)', fontSize: 20, fontFamily: 'Alata, sans-serif' }}>Day High (U$)</span>
-            </div>
-            <span style={{
-              color: 'var(--green)', fontSize: 28, fontFamily: 'Alata, sans-serif', fontWeight: 'bold', margin: '-5px',
-            }}
-            >
-              {state.dayHigh.toFixed(2)}
-            </span>
-          </div>
-          <div className="d-flex flex-column align-items-end">
-            <div style={{
-              borderBottomStyle: 'solid', borderBottomColor: 'var(--light)', height: 30, width: '100%',
-            }}
-            >
-              <span style={{ color: 'var(--light)', fontSize: 20, fontFamily: 'Alata, sans-serif' }}>Day Low (U$)</span>
-            </div>
-            <span style={{
-              color: 'var(--red)', fontSize: 28, fontFamily: 'Alata, sans-serif', fontWeight: 'bold', margin: '-5px',
-            }}
-            >
-              {state.dayLow.toFixed(2)}
-            </span>
-          </div>
-          <div className="d-flex flex-column align-items-end">
-            <div style={{
-              borderBottomStyle: 'solid', borderBottomColor: 'var(--light)', height: 30, width: '100%',
-            }}
-            >
-              <span style={{ color: 'var(--light)', fontSize: 20, fontFamily: 'Alata, sans-serif' }}>Year High  (U$)</span>
-            </div>
-            <span style={{
-              color: 'var(--green)', fontSize: 28, fontFamily: 'Alata, sans-serif', fontWeight: 'bold', margin: '-5px',
-            }}
-            >
-              <strong>{state.yearHigh.toFixed(2)}</strong>
-            </span>
-          </div>
-          <div className="d-flex flex-column align-items-end">
-            <div style={{
-              borderBottomStyle: 'solid', borderBottomColor: 'var(--light)', height: 30, width: '100%',
-            }}
-            >
-              <span style={{ color: 'var(--light)', fontSize: 20, fontFamily: 'Alata, sans-serif' }}>Year Low (U$)</span>
-            </div>
-            <span style={{
-              color: 'var(--red)', fontSize: 28, fontFamily: 'Alata, sans-serif', fontWeight: 'bold', margin: '-5px',
-            }}
-            >
-              <strong>{state.yearLow.toFixed(2)}</strong>
-            </span>
-          </div>
-        </div>
-      </div>
-    </>
-  );
-};
-
-export default Details;
