+++ src/pages/home/Home.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import StockList from '../components/StockList';
-import Navbar from '../components/Navbar';
+import StockList from './components/StockList';
+import Navbar from '../common/Navbar';
 
 const Home = () => (
   <div className="App">
