+++ src/pages/home/components/StockCard.js
